More Semantic HTML
Hi Ragilbuaj, Richard here from that Front End Mentor comment. Thought I'd edit the tags and hopefully make them a bit more semantic for the browser and screen readers and send you a pull request. Hope this helps, and feel free to reject if it's not what you want. All the best:)